backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
yt-dlp>=2024.3.10
google-genai>=1.0.0
google-api-python-client>=2.100.0
pydantic>=2.10.0
redis>=5.0.0
cachetools>=5.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
curl_cffi>=0.7.0
lxml>=4.9.0
recipe-scrapers>=14.50.0